# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mango",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/setup": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Iniciar",
        "operationId": "iniciar_api_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupIniciado"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/setup/confirmar": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Confirmar",
        "operationId": "confirmar_api_setup_confirmar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmarSetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/perfil": {
      "get": {
        "tags": [
          "perfil"
        ],
        "summary": "Obter",
        "operationId": "obter_api_perfil_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilRead"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "perfil"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_perfil_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/perfil/brapi-token": {
      "put": {
        "tags": [
          "perfil"
        ],
        "summary": "Definir Brapi Token",
        "operationId": "definir_brapi_token_api_perfil_brapi_token_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrapiTokenSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "perfil"
        ],
        "summary": "Remover Brapi Token",
        "operationId": "remover_brapi_token_api_perfil_brapi_token_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/itens-pluggy/{item_id}/sincronizar": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Sincronizar Item",
        "operationId": "sincronizar_item_api_itens_pluggy__item_id__sincronizar_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoSyncRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrapiTokenSet": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "BrapiTokenSet",
        "description": "Grava o token brapi (write-only, cifrado em repouso). Nunca é lido de volta."
      },
...
      "ConfirmarSetupRequest": {
        "properties": {
          "setup_ticket": {
            "type": "string",
            "minLength": 1,
            "title": "Setup Ticket"
          },
          "codigo_totp": {
            "type": "string",
            "maxLength": 8,
            "minLength": 6,
            "title": "Codigo Totp"
          }
        },
        "type": "object",
        "required": [
          "setup_ticket",
          "codigo_totp"
        ],
        "title": "ConfirmarSetupRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "accent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "email"
        ],
        "title": "MeRead"
      },
...
      "PerfilRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "data_nascimento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Nascimento"
          },
          "salario_mensal_centavos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salario Mensal Centavos"
          },
          "formacao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formacao"
          },
          "ocupacao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocupacao"
          },
          "accent": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "violeta",
                  "manga",
                  "verde",
                  "azul",
                  "rosa",
                  "teal"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "brapi_token_configurado": {
            "type": "boolean",
            "title": "Brapi Token Configurado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "email",
          "data_nascimento",
          "salario_mensal_centavos",
          "formacao",
          "ocupacao",
          "accent",
          "avatar",
          "brapi_token_configurado"
        ],
        "title": "PerfilRead"
      },
      "PerfilUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "data_nascimento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Nascimento"
          },
          "salario_mensal_centavos": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Salario Mensal Centavos"
          },
          "formacao": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Formacao"
          },
          "ocupacao": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocupacao"
          },
          "accent": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "violeta",
                  "manga",
                  "verde",
                  "azul",
                  "rosa",
                  "teal"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "title": "PerfilUpdate"
      },
      "PluggyCredenciais": {
        "properties": {
          "client_id": {
            "type": "string",
            "minLength": 1,
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "minLength": 1,
            "title": "Client Secret"
          },
          "item_id": {
            "type": "string",
            "minLength": 1,
            "title": "Item Id"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "client_secret",
          "item_id"
        ],
        "title": "PluggyCredenciais"
      },
...
      "ResumoSyncRead": {
        "properties": {
          "itens": {
            "type": "integer",
            "title": "Itens"
          },
          "contas": {
            "type": "integer",
            "title": "Contas"
          },
          "transacoes": {
            "type": "integer",
            "title": "Transacoes"
          },
          "transacoes_novas": {
            "type": "integer",
            "title": "Transacoes Novas"
          },
          "investimentos": {
            "type": "integer",
            "title": "Investimentos"
          }
        },
        "type": "object",
        "required": [
          "itens",
          "contas",
          "transacoes",
          "transacoes_novas",
          "investimentos"
        ],
        "title": "ResumoSyncRead"
      },
...
      "SetupIniciado": {
        "properties": {
          "totp_secret": {
            "type": "string",
            "title": "Totp Secret"
          },
          "totp_provisioning_uri": {
            "type": "string",
            "title": "Totp Provisioning Uri"
          },
          "setup_ticket": {
            "type": "string",
            "title": "Setup Ticket"
          }
        },
        "type": "object",
        "required": [
          "totp_secret",
          "totp_provisioning_uri",
          "setup_ticket"
        ],
        "title": "SetupIniciado",
        "description": "Passo 1: material do TOTP p/ o QR + ticket cifrado (nada foi persistido ainda)."
      },
      "SetupRequest": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "maxLength": 320,
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 8,
            "title": "Senha"
          },
          "pluggy": {
            "$ref": "#/components/schemas/PluggyCredenciais"
          },
          "data_nascimento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Nascimento"
          },
          "salario_mensal": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salario Mensal"
          },
          "formacao": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Formacao"
          },
          "ocupacao": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocupacao"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha",
          "pluggy"
        ],
        "title": "SetupRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}